+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -11,7 +11,7 @@ import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Person;
+import seedu.address.model.types.person.Person;
 
 /**
  * Adds a person to the address book.
