+++ src/main/java/seedu/address/model/types/person/UniquePersonList.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.types.person;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
@@ -8,8 +8,8 @@ import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.person.exceptions.DuplicatePersonException;
-import seedu.address.model.person.exceptions.PersonNotFoundException;
+import seedu.address.model.types.person.exceptions.DuplicatePersonException;
+import seedu.address.model.types.person.exceptions.PersonNotFoundException;
 
 /**
  * A list of persons that enforces uniqueness between its elements and does not allow nulls.
