+++ src/test/java/seedu/address/logic/commands/DeleteCommandTest.java
@@ -17,7 +17,7 @@ import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Person;
+import seedu.address.model.types.person.Person;
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for
