+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -37,12 +37,12 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.AddCommand;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
+import seedu.address.model.types.common.Address;
+import seedu.address.model.types.common.Email;
+import seedu.address.model.types.common.Name;
+import seedu.address.model.types.common.Phone;
+import seedu.address.model.types.person.Person;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddCommandParserTest {
