+++ src/test/java/seedu/address/logic/parser/ParserUtilTest.java
@@ -14,11 +14,11 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
+import seedu.address.model.types.common.Address;
+import seedu.address.model.types.common.Email;
+import seedu.address.model.types.common.Name;
+import seedu.address.model.types.common.Phone;
 
 public class ParserUtilTest {
     private static final String INVALID_NAME = "R@chel";
