+++ src/test/java/seedu/address/model/person/NameTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.types.common.Name;
+
 public class NameTest {
 
     @Test
