+++ src/test/java/seedu/address/model/person/PhoneTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.types.common.Phone;
+
 public class PhoneTest {
 
     @Test
