+++ src/test/java/seedu/address/model/person/UniquePersonListTest.java
@@ -15,8 +15,10 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.person.exceptions.DuplicatePersonException;
-import seedu.address.model.person.exceptions.PersonNotFoundException;
+import seedu.address.model.types.person.Person;
+import seedu.address.model.types.person.UniquePersonList;
+import seedu.address.model.types.person.exceptions.DuplicatePersonException;
+import seedu.address.model.types.person.exceptions.PersonNotFoundException;
 import seedu.address.testutil.PersonBuilder;
 
 public class UniquePersonListTest {
