+++ src/test/java/seedu/address/storage/JsonAdaptedPersonTest.java
@@ -12,10 +12,10 @@ import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Phone;
+import seedu.address.model.types.common.Address;
+import seedu.address.model.types.common.Email;
+import seedu.address.model.types.common.Name;
+import seedu.address.model.types.common.Phone;
 
 public class JsonAdaptedPersonTest {
     private static final String INVALID_NAME = "R@chel";
