+++ src/test/java/seedu/address/testutil/PersonUtil.java
@@ -10,8 +10,8 @@ import java.util.Set;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
-import seedu.address.model.person.Person;
 import seedu.address.model.tag.Tag;
+import seedu.address.model.types.person.Person;
 
 /**
  * A utility class for Person.
